wallpaper + matugen script improvement
- skip running matugen & generating overview bg if same bg is detected
- themechange.sh no longer hardcoded to waypaper bg

+++ private_dot_local/bin/executable_wallswitch.sh
@@ -8,15 +8,19 @@ if [[ ! -f "$1" ]]; then
   exit 1
 fi
 
-if [ "$current_theme" = "'prefer-dark'" ]; then
-  matugen image "$1" -m dark
-elif [ "$current_theme" = "'prefer-light'" ]; then
-  matugen image "$1" -m light
-fi
+if [[ "$1" != "$(readlink -f "$cache_dir/niri/landing/background")" ]]; then
+  if [ "$current_theme" = "'prefer-dark'" ]; then
+    matugen image "$1" -m dark
+  elif [ "$current_theme" = "'prefer-light'" ]; then
+    matugen image "$1" -m light
+  fi
 
-cp -sf "$1" "${cache_dir}"/niri/landing/background
-magick "$1" -blur 20x10 "${blur_img}"
-# waypaper doesn't work properly with multiple swww-daemon instances
-# swww img -n overview --transition-duration 2 -t fade "${blur_img}"
+  cp -sf "$1" "${cache_dir}"/niri/landing/background
+  magick "$1" -blur 20x10 "${blur_img}"
+  # waypaper doesn't work properly with multiple swww-daemon instances
+  # swww img -n overview --transition-duration 2 -t fade "${blur_img}"
+else
+  echo "Same wallpaper found. Skipping matugen & caching..."
+fi
 systemctl --user restart overview-blur.service
 notify-send -e -r 2 -t 2000 "Wallpaper switch successful..." "Current Wallpaper: $(basename "$1")"
